% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_time_stack.R
\name{assign_time_stack}
\alias{assign_time_stack}
\title{Takes a .nc or rasterstack that was generated from openEO and assigns time to the stack.}
\usage{
assign_time_stack(rastStack)
}
\arguments{
\item{rastStack}{terra raster stack from openEO.}
}
\description{
Takes a .nc or rasterstack that was generated from openEO and assigns time to the stack.
}
\examples{
PLACEHOLDER()


}
\keyword{openEO,}
\keyword{stack,}
\keyword{time}
