% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_approximation.R
\name{benchmark_approximation}
\alias{benchmark_approximation}
\title{Temporal remote sensing data often has NA by cloud and other types of masking.
This function approximates linear, spline, HANTS and median normalized difference to asses performance of these methods.
It returns a small data frame with the performance of approximation with Mean Absolute Error (MAE), Root Mean Square Error (RMSE), Bias and R-squared (R2).
Optionally, users can get a smoothScatter for the fit of the approximated against the real values.}
\usage{
benchmark_approximation(
  stack,
  methods,
  HANTS_freq,
  max_gap = 5,
  max_NA_prop = 0.5,
  maxVal = 1,
  minVal = NULL,
  prop_NA = 0.1,
  seed = 43,
  verbose = F,
  ncore = 1,
  figure_path = NULL
)
}
\arguments{
\item{stack}{stack. Stack of rasters that need to be interpolated.}

\item{methods}{character vector. Names of the method to approximate NAs.Available are c("linear","spline","HANTS","ND").}

\item{HANTS_freq}{Integer. Frequency parameter for HANTS interpolation.  Should be the frequency of phenology.}

\item{max_gap}{integer defaults to 4. Number of NA consequetively allowed before returning the original values of the pixel. Applies to HANTS and spline}

\item{max_NA_prop}{numeric defaults to 0.5. Proportion of maximimum NAs allowed per pixel for calculcations. Applies to HANTS and spline}

\item{maxVal}{numeric. Maximum value allowed by spline interpolation. Values higher will become maxVal.}

\item{minVal}{numeric. Minimum value allowed by spline interpolation. Values lower will become minVal.}

\item{prop_NA}{numeric (0-1). Defaults to 0.1. Proportion of non-NA cells to artificially mask}

\item{seed}{integer. defaults to random, but can be set to be reproducable.}

\item{verbose}{logical. True to get extra status updates.}

\item{figure_path}{character. defaults to NULL. Set to path to save a smoothScatter figure between approximate values and true values.}
}
\description{
NOTE: This function can run into std::bad_alloc errors. Reducing terraOptions(mem_frac) tends to help.
Literature:
}
\examples{
PLACEHOLDER()

}
