% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dhm.R
\name{build_dhm}
\alias{build_dhm}
\title{build_dhm; build digital height model}
\usage{
build_dhm(dtm, dsm)
}
\arguments{
\item{dtm}{raster. Digital terrain model}

\item{dsm}{raster. Digital surface model}
}
\description{
This function takes raster digital terrain models and digital surface models to calculate the digital height model. This can be understood as the height from the ground to a surface above it. NA values from the digital terrain model will be bilinearly interpolated based on a 3x3 grid. NOTE: this function currently only works with AHN data.
}
\examples{
build_dhm()
}
\keyword{Height}
\keyword{model}
