% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_height.R
\name{classify_height}
\alias{classify_height}
\title{Classifying height and proportionalise to template}
\usage{
classify_height(clean_dhm, template)
}
\arguments{
\item{clean_dhm}{A raster representing a digital height model.}

\item{template}{A raster (of coarser resolution) that can be used to calculate proportions per height class.}
}
\description{
This function takes a raster representing height and a template to generate three rasters of proportional grass, shrub and trees.
First, This function classifies height as grass (-0.1,1), shrub (1 - 5) or tree (5 +).
After this function calculates the proportion of the height class through average resampling
NOTE:
Literature: CUGIC: The Consolidated Urban Green Infrastructure Classification for assessing ecosystem services and biodiversity, https://doi.org/10.1016/j.landurbplan.2023.104726
}
\examples{
classify_height()

}
\keyword{classification,}
\keyword{height,}
\keyword{proportion}
