% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cugicify.R
\name{cugicify}
\alias{cugicify}
\title{Make the Consolidated Urban Green Infrastructure Classification}
\usage{
cugicify(city_name, date, veg_coverage, veg_structure, lulc)
}
\arguments{
\item{city_name}{character. Name of the city for the analysis}

\item{date}{character. dd-mm-yyyy of the analysis}

\item{veg_coverage}{raster. Fractional coverage by vegetation or absolute coverage by vegetation.}

\item{veg_structure}{spatialObject. Indicating vegetation height. Either raster with numeric values OR vector with values ("grass","shrub","wooded").}

\item{lulc}{vector. Contains LULC that relates to the lulc_info data table associated with the package}
}
\description{
This function take in a ndvi map and classified height maps and combines to map the CUGIC.
NOTE:
Literature: CUGIC: The Consolidated Urban Green Infrastructure Classification for assessing ecosystem services and biodiversity. https://doi.org/10.1016/j.landurbplan.2023.104726
}
\examples{
make_cugic()

}
\keyword{classification,}
\keyword{height,}
\keyword{proportion}
