% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_gain_loss.R
\name{diff_gain_loss}
\alias{diff_gain_loss}
\title{Map the gains and losses over a time series}
\usage{
diff_gain_loss(raster_stack)
}
\arguments{
\item{raster_stack}{A raster stack representing a time series}
}
\description{
This function takes a raster stack, representing a time-series, to calculate the loss and gains per step in the time series.
It does so by calculating the difference per time step and creates a stack of differences, gains or losses in separate stacks.
The function returns 2 additional rasters which represent the cumulative gains and losses, solely based on the gains and losses stack.
NOTE:
Literature:
}
\examples{
diff_gain_loss()

}
\keyword{classification,}
\keyword{height,}
\keyword{proportion}
