% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ahn_files.r
\name{download_ahn_file}
\alias{download_ahn_file}
\title{Download AHN files}
\usage{
download_ahn_file(
  versie,
  type,
  blad,
  base_dir,
  base_url = "https://fsn1.your-objectstorage.com/hwh-ahn",
  remove_zip = TRUE
)
}
\arguments{
\item{versie}{A character value; being AHN3, AHN4 or AHN5}

\item{type}{A character value; being either DSM_50cm or DTM_50cm}

\item{blad}{A character value; Indicating the tile you want to download}

\item{base_dir}{A character value; path to folder where downloads need to go}

\item{base_url}{A character value; Defaults to https://fsn1.your-objectstorage.com/hwh-ahn}

\item{remove_zip}{TRUE or FALSE; to remove the downloaded .zip}
}
\description{
This function is able to download both DTM and DSM of Actueel Hoogte Bestand Nederland (AHN) version 3-5.
The function also is able to unzips the downloads and removes the .zip.
NOTE:
Literature:
}
\examples{
download_ahn_file()

}
\keyword{download,}
\keyword{height}
\keyword{netherlands,}
\keyword{request,}
