% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_layers.R
\name{fill_missing_layers}
\alias{fill_missing_layers}
\title{Fills in the temporally missing layers in a raster stack on monthly basis.}
\usage{
fill_missing_layers(r, interval = "month")
}
\arguments{
\item{r}{A terra SpatRaster object with time information (\code{terra::time(r)}).}
}
\value{
The stack with NA for missing layers.
#' PLACEHOLDER()
}
\description{
Fills in the temporally missing layers in a raster stack on monthly basis.
}
\examples{
placeholder
}
