% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest_chr_match.R
\name{find_closest_chr_match}
\alias{find_closest_chr_match}
\title{Find closest character match}
\usage{
find_closest_chr_match(max.dist = 0.9, stringX = NULL, stringY = NULL)
}
\arguments{
\item{max.dist}{Numeric value; Indicating maximum distance allowed between X and the result. Value must be between 0 - 1.
0 needs the result to be the same as X. 1 Allows for maximum difference between X and the result.}

\item{stringX}{Character value; The character value you are looking for or similar in a list.}

\item{stringY}{Character values; A vector containing the multiple character values of which you want to find the most similar to stringX.}
}
\description{
This function is an iterative wrapper around agrep. It matches strings and does so until a best answer is found.
NOTE:
Literature:
}
\examples{
find_closest_chr_match()

}
\keyword{similarity}
\keyword{words,}
