% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_ahn.r
\name{get_data_ahn}
\alias{get_data_ahn}
\title{Download AHN files}
\usage{
get_data_ahn(date, city_name)
}
\arguments{
\item{date}{Character value; dd-mm-yyyy}

\item{city_name}{Character value; city and country ideally. e.g. "Alkmaar, the Netherlands"}
}
\description{
This function downloads both DTM and DSM of Actueel Hoogte Bestand Nederland (AHN) version 3-5 to make a digital height model.
NOTE:
Literature:
}
\examples{
download_ahn_file()

}
\keyword{download,}
\keyword{height}
\keyword{netherlands,}
\keyword{request,}
