% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_openeo_data.R
\name{get_data_openeo}
\alias{get_data_openeo}
\title{API request for GEE data}
\usage{
get_data_openeo(
  aoi,
  city_name,
  startdate,
  enddate,
  satellite,
  indicator,
  method,
  cloud_threshold = 80
)
}
\arguments{
\item{aoi}{an object with an extent}

\item{city_name}{character. Name of the city for the data retrieval}

\item{startdate}{date. Start of the requesting images}

\item{enddate}{date. End of the requesting images}

\item{indicator}{Character value; NDVI (future; EVI or LST)}

\item{method}{Character value. max, 90th, median, monthly_max, monthly_90th, monthly_median or none (for all images). 90th indicates the 90th percentile of a pixel value over the time period. Note: none will retrieve massive amounts of images.}

\item{cloud_threshold}{numeric. Standard at 50.}

\item{sattelite}{character value. Sentinel or (future; Landsat)}
}
\description{
This function works with openeo to request data via API requests to R.
}
\details{
NOTE: https://openeo.dataspace.copernicus.eu/ For Sentinel2 data is harmonizes the values before and after baseline shift occuring in 2025.
Literature:
}
\examples{
get_openeo_data()

}
\keyword{API,}
\keyword{Indicators,}
\keyword{Remote}
\keyword{Sensing,}
\keyword{openEO}
