% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lulc_ohsome.R
\name{get_lulc_ohsome}
\alias{get_lulc_ohsome}
\title{Retrieves historic LULC data from Ohsome}
\usage{
get_lulc_ohsome(aoi, keyval_df, year, filename = NULL)
}
\arguments{
\item{aoi}{SpatRaster, raster, sf or character value. Is the area of interest}

\item{keyval_df}{dataframe or matrix with 2 columns; 1) key 2) value. This corresponds to Ohsome datastructure.}

\item{year}{numeric value. Indicates the year of the data you request.}

\item{filename}{character value. Path to folder where the file should be saved as a .gpkg}
}
\description{
This function is a wrapper for ohsome_elements_geometry to retrieve historic LULC data from Ohsome.
It improves it by making the request robust if it fails for some reason, omit extra data, combines poly and multipolygons, and allows for direct saving to disk.
NOTE:
Literature:
}
\examples{
PLACEHOLDER()

}
\keyword{Land-cover,}
\keyword{Land-use,}
\keyword{OSM,}
\keyword{Ohsome,}
\keyword{historic}
