% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_municipal_bbox.R
\name{get_municipal_bbox}
\alias{get_municipal_bbox}
\title{Retrieve municipal bbox as sf}
\usage{
get_municipal_bbox(city_name)
}
\arguments{
\item{city_name}{character value. name of a city to retrieve the bbox for.}
}
\description{
This function is a wrapper around getbb and returns an sf object saved as an .rds on drive.
NOTE:
Literature:
}
\examples{
PLACEHOLDER()

}
