% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANTS_pixel.R
\name{HANTS_pixel}
\alias{HANTS_pixel}
\title{HANTS curve for a vector or 1 pixel in a raster stack. Requires at least half of pixels to have values.}
\usage{
HANTS_pixel(
  y,
  freq = 1,
  max_iter = 3,
  tolerance = 0.1,
  max_gap = 5,
  max_NA_prop = 0.5
)
}
\arguments{
\item{y}{numeric vector. Time series for harmonic interpolation following HANTS algoritm.}

\item{freq}{integer. Number of harmonic frequencies to fit}

\item{max_iter}{integer. Max number of iterations for outlier rejection.}

\item{tolerance}{Numeric. Threshold for outlier rejection as absolute numeric value. Default is 0.10. In Roerink et al., (2000), the value used is 0.05.}

\item{max_gap}{Numeric (defaults to 5). The number of maximum NA gap allowed before the values are directly returned}

\item{max_NA_prop}{Numeric (defaults to 0.5). The proportion of NA that is allows in the vector for interpolation.}
}
\description{
NOTE:
Literature: Roerink, G. J., Menenti, M., & Verhoef, W. (2000). Reconstructing cloudfree NDVI composites using Fourier analysis of time series. International Journal of Remote Sensing, 21(9), 1911–1917. https://doi.org/10.1080/014311600209814
}
\examples{
PLACEHOLDER()

}
