% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANTS_stack.R
\name{HANTS_stack}
\alias{HANTS_stack}
\title{HANTS for a spatRaster stack. Frequency should be the frequencies expected in the dataset. E.g. annual cycle would be freq = 1. Function returns an identical stack with interpolated values at the NAs.}
\usage{
HANTS_stack(
  stack,
  cores = 1,
  freq = 1,
  max_iter = 3,
  tolerance = 0.1,
  max_gap = 5,
  max_NA_prop = 0.5
)
}
\arguments{
\item{stack}{stack raster Terra. Stack of rasters that need to be interpolated.}
}
\description{
NOTE:
Literature: Roerink, G. J., Menenti, M., & Verhoef, W. (2000). Reconstructing cloudfree NDVI composites using Fourier analysis of time series. International Journal of Remote Sensing, 21(9), 1911–1917. https://doi.org/10.1080/014311600209814
}
\examples{
PLACEHOLDER()

}
