% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactivebbox_DEPRECATED.R
\name{interactiveBbox}
\alias{interactiveBbox}
\title{Helper function that allows for interactive selection of municipalities}
\usage{
interactiveBbox(city_name, sleep = 1)
}
\arguments{
\item{city_name}{character value. Name of a city}

\item{sleep}{numeric value. default = 1. This helps the function to be stable as mapview works via the viewer pane.}
}
\description{
This function takes municipal borders from OSM and then allows the user to select another municipal border via Mapview if the initial border by the function is wrong.
}
\details{
NOTE:
Literature:
}
\examples{
interactiveBbox()

}
\keyword{classification,}
\keyword{height,}
\keyword{proportion}
