% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_CUGIC.R
\name{make_cugic}
\alias{make_cugic}
\title{Make the Consolidated Urban Green Infrastructure Classification}
\usage{
make_cugic(
  ndvi_stack,
  ndvi_layer_index = 1,
  classified_height,
  ndvi_min = 0.25,
  out_fcv = NULL,
  out_cugic = NULL
)
}
\arguments{
\item{ndvi_stack}{rasterStack. Containing ndvi rasters}

\item{ndvi_layer_index}{numeric value. Indicating for which layer the function uses}

\item{classified_height}{rasterStack. Containing 3 layers corresponding to grass, shrub and tree proportions.}

\item{ndvi_min}{numeric value. Cut-off below which are considered non-vegetation}

\item{out_fcv}{character value. Path for saving fractional coverage of vegetation}

\item{out_cugic}{chracter value. Path for saving CUGIC output}
}
\description{
This function take in a ndvi map and classified height maps and combines to map the CUGIC.
NOTE:
Literature: CUGIC: The Consolidated Urban Green Infrastructure Classification for assessing ecosystem services and biodiversity. https://doi.org/10.1016/j.landurbplan.2023.104726
}
\examples{
make_cugic()

}
\keyword{classification,}
\keyword{height,}
\keyword{proportion}
