% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_osm_polygons.R
\name{merge_osm_polygons}
\alias{merge_osm_polygons}
\title{Helper function to make OSM/Ohsome output consistent}
\usage{
merge_osm_polygons(osmdata_obj)
}
\arguments{
\item{osmdata_obj}{osm_output. The output from OSM API request.}
}
\description{
This function takes POLYGON and MULTIPOLYGON and unifies it to polygons.
NOTE:
Literature:
}
\examples{
merge_osm_polygons()
}
\keyword{classification,}
\keyword{height,}
\keyword{proportion}
