% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_and_mask.R
\name{mosaic_and_mask}
\alias{mosaic_and_mask}
\title{Crops list of rasters to AOI}
\usage{
mosaic_and_mask(raster_dir, aoi_file, out_file = "mosaic_masked.tif")
}
\arguments{
\item{raster_dir}{path to folder with raster files}

\item{aoi_file}{path to file that canbe loaded by vect() from Terra.}

\item{out_file}{Optional. Path to place where to put the output.}
}
\description{
This function takes a list of rasters and check if the extent overlaps the AOI. If it does it mosaics them together and the crops them.
NOTE:
Literature:
}
\examples{
mosaic_and_mask()

}
\keyword{classification,}
\keyword{height,}
\keyword{proportion}
