% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndviToDensity.R
\name{ndviToDensity}
\alias{ndviToDensity}
\title{Transform NDVI values into both Fractional vegetation coverage and then into absolute coverage into m2.}
\usage{
ndviToDensity(NDVI_raster, ndvi_min = 0.25, m2 = F)
}
\arguments{
\item{NDVI_raster}{raster. This files contains the values of NDVI.}
}
\description{
NOTE:
Literature: Carlson, T. N., & Ripley, D. A. (1997). On the relation between NDVI, fractional vegetation cover, and leaf area index. Remote Sensing of Environment, 62(3), 241–252. https://doi.org/10.1016/S0034-4257(97)00104-1
}
\examples{
PLACEHOLDER()

}
