% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_to_local_utm.R
\name{project_to_local_utm}
\alias{project_to_local_utm}
\title{This function is made to check where the geospatial data is located on the globe and projects the local UTM to have a good approximation of meters for downstream analysis.}
\usage{
project_to_local_utm(object)
}
\arguments{
\item{object}{Geospatial object. This can be sf, raster or terra object.}
}
\description{
NOTE:
Literature:
}
\examples{
PLACEHOLDER()

}
