% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_api_request.R
\name{robust_api_request}
\alias{robust_api_request}
\title{Retry failed API requests}
\usage{
robust_api_request(osm_function, ..., max_attempts = 6, wait_time = 10)
}
\arguments{
\item{osm_function}{function for sending API request}

\item{...}{arguments for osm_function}

\item{max_attempts}{numeric value. Maximum number of attempts}

\item{wait_time}{numeric value. Wait between attempts.}
}
\description{
This function retries failed API requests with increasing long wait periods.
Wait period is calculated as attempt * wait time.
NOTE:
Literature:
}
\examples{
robust_api_request()

}
\keyword{classification,}
\keyword{height,}
\keyword{proportion}
