% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_rasters_equalise_extent.R
\name{stack_rasters_equalise_extent}
\alias{stack_rasters_equalise_extent}
\title{Stacks and equalises rasters}
\usage{
stack_rasters_equalise_extent(rasterFiles)
}
\arguments{
\item{rasterFiles}{path to raster files.}
}
\description{
This function takes a list of raster file path and equalises their extent. After it stacks the rasters.
NOTE:
Literature:
}
\examples{
mask_buildings_ohsome_robust()

}
\keyword{classification,}
\keyword{height,}
\keyword{proportion}
