% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_assessment_streetview.R
\name{streetview_assessment_green}
\alias{streetview_assessment_green}
\title{NOTE:
Literature:}
\usage{
streetview_assessment_green(raster)
}
\arguments{
\item{raster}{raster.}
}
\description{
NOTE:
Literature:
}
\examples{
PLACEHOLDER()

}
