% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_spatial_approximation.R
\name{temporal_spatial_approximation}
\alias{temporal_spatial_approximation}
\title{Temporal interpolation based on median normalized difference between two layers. This approach assumes both only positive values and constant effect across all cells.}
\usage{
temporal_spatial_approximation(stack)
}
\arguments{
\item{stack}{spat raster stack. Stack of rasters to be interpolated}
}
\description{
NOTE:
Literature:
}
\examples{
PLACEHOLDER()

}
