% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_osm_cityname_query.R
\name{test_osm_cityname_query}
\alias{test_osm_cityname_query}
\title{Test city name(s) in OSM}
\usage{
test_osm_cityname_query(cities, map = F)
}
\arguments{
\item{cities}{character vector. Value containing city names}

\item{map}{TRUE or FALSE. If map with points of the cities should be shown}
}
\description{
This function uses geo() to test city name(s) and then checks if the results are OK and which are pot missing or wrong.
NOTE:
Literature:
}
\examples{
diff_gain_loss()

}
\keyword{classification,}
\keyword{height,}
\keyword{proportion}
