% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VItoDensity.R, R/ndviToDensity.R
\name{VItoDensity}
\alias{VItoDensity}
\title{Transform Vegetation Index (VI) values into Fractional vegetation coverage and optionally into absolute coverage into m2.
This function work on dynamic approach by validating the VI as a classifier for green and grey by LULC maps.
NDVI is the standard VI, however EVI and MSAVI typically perform better. The function searches for files in the MUST_downloaded_data folder in the city and VI subfolder for a file of the correct year and ending in median.tif
The standard method the VI used is annual median. Max and 90th percentile are not recommend nor verified nor possible in this function.}
\usage{
VItoDensity(
  city_name,
  date,
  VI = NULL,
  lulc = NULL,
  ndvi_min = 0.25,
  m2 = F,
  pathFigure = NULL
)

VItoDensity(
  city_name,
  date,
  VI = NULL,
  lulc = NULL,
  ndvi_min = 0.25,
  m2 = F,
  pathFigure = NULL
)
}
\arguments{
\item{city_name}{character. Relates to the place where to look for NDVI.}

\item{date}{character. The year of the NDVI (and optionally LULC-map) to consider. Works only with YYYY-MM-DD format.}

\item{VI}{character or terra rast. Character needs to be one of the follow: c("NDVI", "EVI", "MSAVI") and will retrieve this file. If a rast is provided this will be used as VI input.}

\item{m2}{logical. standard is F. When set to T, an extra multiplication is done by cell size to estimate the m2 covered by vegetation.}

\item{pathFigure}{character. Where to save the resulting image.}

\item{VI_min}{numeric OR "Dynamic". Numeric (default = 0.25) denotes the minimal threshold value of NDVI to be considered 0\% vegetation coverage. The dynamic option requires LULC to be downloaded and draws random point in LULC that is and isn't green to determine a threshold. This also produces a plot and some extra information.}

\item{figure}{character. Where to save the resulting image, defaults to results folder}

\item{maskLayer}{spatVector. Object to mask the VI raster by.}

\item{save}{logical. Standard is FALSE. When set to true the VItoDensity map is saved to the results folder. You may also set a path with a character value.}

\item{NDVI_min}{numeric OR "Dynamic". Numeric (default = 0.25) denotes the minimal threshold value of NDVI to be considered 0\% vegetation coverage. The dynamic option requires LULC to be downloaded and draws random point in LULC that is and isn't green to determine a threshold. This also produces a plot and some extra information.}
}
\description{
NOTE:
Literature: Carlson, T. N., & Ripley, D. A. (1997). On the relation between NDVI, fractional vegetation cover, and leaf area index. Remote Sensing of Environment, 62(3), 241–252. https://doi.org/10.1016/S0034-4257(97)00104-1

NOTE:
Literature: Carlson, T. N., & Ripley, D. A. (1997). On the relation between NDVI, fractional vegetation cover, and leaf area index. Remote Sensing of Environment, 62(3), 241–252. https://doi.org/10.1016/S0034-4257(97)00104-1
}
\examples{
PLACEHOLDER()

PLACEHOLDER()

}
