% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_GEE.R
\name{get_GEE_data}
\alias{get_GEE_data}
\title{API request for GEE data}
\usage{
get_GEE_data(
  aoi,
  start_date,
  end_date,
  sattelite = c("Sentinel", "Landsat"),
  method = c("NDVI", "EVI", "LST"),
  indicator = c("median", "max"),
  temporal_resolution = c("none", "annual", "monthly"),
  CLOUD_FILTER = 60,
  quiet = T,
  lazy = T,
  container = NULL,
  via = NULL,
  dsn = NULL,
  maxPixels = 1e+10,
  scale = 1000
)
}
\arguments{
\item{aoi}{ee.object}

\item{start_date}{date. Start of the requesting images}

\item{end_date}{date. End of the requesting images}

\item{sattelite}{character value. Sentinel or Landsat}

\item{method}{Character value; NDVI, EVI or LST}

\item{indicator}{character value; Median or Max}

\item{temporal_resolution}{character value; Currently only annual}

\item{CLOUD_FILTER}{numeric value; between 0 - 100}

\item{quiet}{TRUE or FALSE.}

\item{lazy}{TRUE or FALSE, to send the request and not wait for it to finish}

\item{container}{= character value. Where to store}

\item{via}{character value. where to store}

\item{dsn}{character value. How to call the file}

\item{maxPixels}{numeric value. maximum pixels in your request}

\item{scale}{numeric value. resolution for data requested.}
}
\description{
This function is a wrapper around the RGEE function ee_as_rast and allows the user to also switch between some basic arguments for rapid data requests.
}
\details{
NOTE:
Literature:
}
\examples{
get_GEE_data()

}
\keyword{API,}
\keyword{GEE}
\keyword{Indicators,}
\keyword{Remote}
\keyword{Sensing,}
