% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_openeo_timeseries_WIP.R
\name{get_timeseries_openeo}
\alias{get_timeseries_openeo}
\title{Retrieve Time Series Image Collection from openEO}
\usage{
get_timeseries_openeo(
  aoi,
  city_name,
  startdate,
  enddate,
  satellite,
  indicator,
  cloud_threshold = 50
)
}
\arguments{
\item{aoi}{an object with an extent (e.g., sf object)}

\item{city_name}{character. Name of the city for file naming}

\item{startdate}{date/character. Start date (YYYY-MM-DD)}

\item{enddate}{date/character. End date (YYYY-MM-DD)}

\item{satellite}{character. "sentinel2" or "landsat8"}

\item{indicator}{character. "NDVI", "EVI", "MSAVI", "LST", or "NDWI"}

\item{cloud_threshold}{numeric. Max cloud cover percentage (0-100)}
}
\description{
This function requests all available images within a time frame without temporal aggregation.
}
