% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urban_border.R
\name{get_urban_border}
\alias{get_urban_border}
\title{Define urban border}
\usage{
get_urban_border(
  city_name = NULL,
  date = NULL,
  bufferSize = 500,
  thresholdUrban = 0.5,
  gap_bridge_distance = 1000,
  gap_filling = T,
  LULC = NULL,
  save = T
)
}
\arguments{
\item{city_name}{character value. Name of a city. This should correspond exactly to the name used in other functions of the package.}

\item{date}{character value. Which date/time to retrieve the snapshot from Ohsome}

\item{bufferSize}{numeric value. Defaults to 500, user input needs to be divisible by 10. This indicate meters size of the circular moving average window to define urban border.}

\item{thresholdUrban}{numeric proportional value. Between 0-1 (defaults to 0.5). This value indicates the threshold proportion for the moving window to be considered urban. i.e. 0.5 will classify a cell as urban when 50\% of the cells from the buffer are classified as urban.}

\item{gap_bridge_distance}{numeric value. Defaults to 1000, user input needs to be divisible by 10. Indicating meters, that borders will grow and shrink with the intend to close gaps. The basic value of this is 1000 (1km), which relates to approx 15min walking. This makes strips of lands classified as non-urban in the city (parks, nature reserves, etc.) that span 2 km urban, but does not affect anything else.}

\item{gap_filling}{Logical. TRUE indicates to fill gaps within the urban border after all other operations. This may help with large parks or non-urban classified land.}

\item{LULC}{.tif. Map from Terra having LULC. The following LULC cats are recoginized as non-urban c("agriculture", "water", "river", "canal", "stream", "sea", "natural", "Green buffer zones").}

\item{save}{FALSE or path. Where to save the define urban border. Defaults to T to save border in the standard folder. F to no save. define path (character) to  place to in particular folder.}
}
\description{
This function take the LULC file, creates a binary raster of urban environment through excluding green LULC. After it fill gaps between urban cell up to 1km and then define the urban border as 50\% of space having a urban LULC in either (default) 500m, 300m or 1000m moving window. NOTE: 1000m will be slow, and can crash, for larger cities.
NOTE:
Literature:
}
\examples{
PLACEHOLDER()

}
\keyword{border}
\keyword{urban,}
