% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urban_border.R
\name{get_urban_border}
\alias{get_urban_border}
\title{Calculate urban border}
\usage{
get_urban_border(
  city_name,
  historic = F,
  date = NULL,
  interactive = F,
  dbscan = T,
  eps = 1000,
  minPts = 100,
  plot = F
)
}
\arguments{
\item{city_name}{character value. Name of a city}

\item{historic}{TRUE or FALSE. Whether to retrieve current data from OpenStreetMaps or Ohsome for historic data}

\item{date}{character value. Which date/time to retrieve the snapshot from Ohsome}

\item{interactive}{TRUE or FALSE. Allows user to selct municipal borders through Mapview. This is not recommended, but can help with problems if this function returns borders for the wrong city.}

\item{dbscan}{TRUE or FALSE. Use Density-based spatial clustering of applications with noise (DBSCAN) to find urban clusters.}

\item{eps}{Distance to be considered}

\item{minPts}{Minimum points for consideration of cluster}

\item{plot}{legacy. Can be set to TRUE}
}
\description{
This function retrieves the municipal border and land-uses from OpenStreetMaps. After it transforms the land-use polygons into points
every 100m^2. These points are put in the DBSCAN algorithm to cluster points. After all clusters in the largest 90\% are considered urban.
Finally, a concave hull is drawn around the remaining urban points which represents an estimate of the urban area of interest (AOI)
NOTE:
Literature:
}
\examples{
PLACEHOLDER()

}
\keyword{DBSCAN}
\keyword{border,}
\keyword{urban,}
