% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urban_border.R
\name{get_urban_border}
\alias{get_urban_border}
\title{Define urban border}
\usage{
get_urban_border(
  city_name = NULL,
  date = NULL,
  bufferSize = NULL,
  LULC = NULL,
  save = T
)
}
\arguments{
\item{city_name}{character value. Name of a city. This should correspond exactly to the name used in other functions of the package.}

\item{date}{character value. Which date/time to retrieve the snapshot from Ohsome}

\item{bufferSize}{numeric value. Defaults to 500. Either, or combination, of 300, 500 or 1000. These indicate meters size of the moving average window to define urban border.}

\item{LULC}{.tif. Map from Terra having LULC. The following LULC cats are recoginized as non-urban c("agriculture", "water", "river", "canal", "stream", "sea", "natural", "Green buffer zones").}

\item{save}{FALSE or path. Where to save the define urban border. Defaults to T to save border in the standard folder. F to no save. define path (character) to  place to in particular folder.}
}
\description{
This function take the LULC file, creates a binary raster of urban environment through excluding green LULC. After it fill gaps between urban cell up to 1km and then define the urban border as 50\% of space having a urban LULC in either (default) 500m, 300m or 1000m moving window. NOTE: 1000m will be slow, and can crash, for larger cities.
NOTE:
Literature:
}
\examples{
PLACEHOLDER()

}
\keyword{border}
\keyword{urban,}
