% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANTS_pixel.R
\name{HANTS_pixel}
\alias{HANTS_pixel}
\title{HANTS curve for 1 pixel}
\usage{
HANTS_pixel(y, freq = 1, max_iter = 3, tolerance = 0.1)
}
\arguments{
\item{y}{numeric vector. Time series for harmonic interpolation following HANTS algoritm.}

\item{freq}{integer. Number of harmonic frequencies to fit}

\item{max_iter}{integer. Max number of iterations for outlier rejection.}

\item{tolerance}{Numeric. Threshold for outlier rejection as absolute numeric value. Default is 0.10. In Roerink et al., (2000), the value used is 0.05.}
}
\description{
NOTE:
Literature: Roerink, G. J., Menenti, M., & Verhoef, W. (2000). Reconstructing cloudfree NDVI composites using Fourier analysis of time series. International Journal of Remote Sensing, 21(9), 1911–1917. https://doi.org/10.1080/014311600209814
}
\examples{
PLACEHOLDER()

}
