% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HANTS_stack.R
\name{HANTS_stack}
\alias{HANTS_stack}
\title{HANTS for a spatRaster stack.}
\usage{
HANTS_stack(stack, cores = 1, freq = 1, max_iter = 3, tolerance = 0.1)
}
\arguments{
\item{stack}{stack raster Terra. Stack of rasters that need to be interpolated.}
}
\description{
NOTE:
Literature: Roerink, G. J., Menenti, M., & Verhoef, W. (2000). Reconstructing cloudfree NDVI composites using Fourier analysis of time series. International Journal of Remote Sensing, 21(9), 1911–1917. https://doi.org/10.1080/014311600209814
}
\examples{
PLACEHOLDER()

}
