% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_duplicate_drive_folders.R
\name{merge_duplicate_drive_folders}
\alias{merge_duplicate_drive_folders}
\title{Check duplicate drive folder and merge them}
\usage{
merge_duplicate_drive_folders(folder_name)
}
\arguments{
\item{folder_name}{character value. Name of the folder you want to be checked for duplication and merged.}
}
\description{
This function checks duplicate drive folders and merges them. WARNING: This will remove empty folder (i.e. folder that are having this name. Use with caution)
NOTE:
Literature:
}
\examples{
merge_duplicate_drive_folders()

}
\keyword{classification,}
\keyword{height,}
\keyword{proportion}
