% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_download_check.R
\name{pre_download_check}
\alias{pre_download_check}
\title{Checks if files are in cache and retrieves if ther}
\usage{
pre_download_check(city_name, object, indicator = NULL)
}
\arguments{
\item{city_name}{folder it will look for}

\item{object}{file to look for}
}
\description{
Checks if files are in cache and retrieves if ther
}
\examples{
PLACEHOLDER()


}
\keyword{Land-cover,}
\keyword{Land-use,}
\keyword{OSM,}
\keyword{Ohsome,}
\keyword{historic}
