% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_download_check.R
\name{pre_download_check}
\alias{pre_download_check}
\title{Checks if files are in cache and retrieves if present.}
\usage{
pre_download_check(city_name, object, subfolder = NULL)
}
\arguments{
\item{city_name}{folder to check in the MUST_downloaded_data folder.}

\item{object}{file to retrieve}

\item{subfolder}{sub-folder to look for the object}
}
\description{
Checks if files are in cache and retrieves if present.
}
\examples{
PLACEHOLDER()


}
\keyword{Land-cover,}
\keyword{Land-use,}
\keyword{OSM,}
\keyword{Ohsome,}
\keyword{historic}
