% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_pixel.R
\name{spline_pixel}
\alias{spline_pixel}
\title{Spline interpolation based on equi-distant time-series}
\usage{
spline_pixel(y)
}
\arguments{
\item{y}{numeric vector. Time series for spline interpolation}
}
\description{
NOTE:
Literature:
}
\examples{
PLACEHOLDER()

}
