% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_structure_via_lulc.R
\name{veg_structure_via_lulc}
\alias{veg_structure_via_lulc}
\title{Generates a classified height map based on LULC and tree points from OSM.}
\usage{
veg_structure_via_lulc(lulc_map)
}
\arguments{
\item{city_name}{character. Name of city and country.}

\item{date}{character. DD-MM-YYYY for date to retrieve data from}
}
\description{
NOTE:
Literature:
}
\examples{
PLACEHOLDER()

}
\keyword{Land-cover,}
\keyword{Land-use,}
\keyword{OSM,}
\keyword{Ohsome,}
\keyword{historic}
