% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VItoDensity.R
\name{VItoDensity}
\alias{VItoDensity}
\title{Transform Vegetation Index (VI) values into Fractional vegetation coverage and optionally into absolute coverage into m2.
This function provides two options to rescale VI into Fractional vegetation coverage.
Either provide a threshold for the VI_min argument or the character value "dynamic"" to use green and grey from LULC maps to determine a threshold via a classification problem.
Threshold by dynamic classification is calculated using a Weighted Youden threshold with sensitivity-weight 0.75 and specificity-weight of 0.25
The dynamic threshold argument also clamps all values below 0, to make the thresholding more reliable in case of odd patterns or data.
The function returns either a raster or stack if save_fcv used with multiple values.}
\usage{
VItoDensity(
  city_name,
  date = NULL,
  VI,
  lulc = NULL,
  VI_min = 0.25,
  weight = 0.75,
  maskLayer = NULL,
  m2 = F,
  figure = F,
  save_fcv = NULL,
  save_path = NULL
)
}
\arguments{
\item{city_name}{character. Relates to the place where to look for data.}

\item{date}{character. The year of the VI (and optionally LULC-map) to consider. Works only with YYYY-MM-DD format. or numeric year as YYYY format.}

\item{VI}{character or terra rast. Character needs to be one of the follow: c("NDVI", "EVI", "MSAVI") and will retrieve this file. If a rast is provided this will be used as VI input.}

\item{VI_min}{numeric OR "Dynamic". Numeric (default = 0.25) denotes the minimal threshold value of NDVI to be considered 0\% vegetation coverage. The dynamic option requires LULC to be downloaded and draws random point in LULC that is and isn't green to determine a threshold. This also produces a plot and some extra information. NOTE: The dynamic approach only takes 1 raster layer. Providing a numeric value allows for the processing of a full Stack.}

\item{weight}{numeric. Defaults to 0.75. Weight to be assigned to the Weighted Youden Index.}

\item{maskLayer}{spatVector. Object to mask the VI raster by.}

\item{m2}{logical. standard is F. When set to T, an extra multiplication is done by cell size to estimate the m2 covered by vegetation.}

\item{figure}{logical or path. Defaults to FALSE to not save a figure. TRUE save figure in the results folder where it shows A) distributions of values for green, grey and blue classes and the calculated threshold and B) Optimization curves with specificity, sensitivity and curves for (weighted) Youden Index. Alternatively, a path can be given to store the figure there.}

\item{save_fcv}{character vector. default is NULL. Provide one or more of the following to scale the VI to that threshold c("95_sens","w_youden","youden"). Raster, or stack, is saved to results folder.}

\item{save_path}{character vector. Defaults to the results folder. Users may specify path to store the fcv raster or stack in a different location or way.}
}
\description{
NOTE:
Literature: Carlson, T. N., & Ripley, D. A. (1997). On the relation between NDVI, fractional vegetation cover, and leaf area index. Remote Sensing of Environment, 62(3), 241–252. https://doi.org/10.1016/S0034-4257(97)00104-1; Li, D., Shen, F., Yin, Y., Peng, J., & Chen, P. (2013). Weighted Youden index and its two-independent-sample comparison based on weighted sensitivity and specificity. Chinese Medical Journal, 126(6), 1150–1154. https://doi.org/10.3760/cma.j.issn.0366-6999.20123102
}
\examples{
PLACEHOLDER()

}
